pyproject.toml
[project]
name = "blog-quent-in"
version = "0.1.0"
description = "Scripts for blog.quent.in Hugo site"
requires-python = ">=3.13"
dependencies = [
    "google-genai",
    "pillow",
    "python-frontmatter",
]

[project.scripts]
new-post = "scripts.new_post:main"
generate-covers = "scripts.generate_covers:main"
